+++ engine/include/sf_containers/traits.hpp
@@ -20,12 +20,12 @@ struct ReallocReturnHandle {
 
 template<typename A, typename T = void, typename ...Args>
 concept AllocatorTrait = requires(A a, Args... args) {
-    { a.allocate(std::declval<u32>(), std::declval<u16>()) } -> std::same_as<T*>;
-    { a.allocate_handle(std::declval<u32>(), std::declval<u16>()) } -> std::same_as<usize>;
+    { a.allocate(std::declval<usize>(), std::declval<u16>()) } -> std::same_as<T*>;
+    { a.allocate_handle(std::declval<usize>(), std::declval<u16>()) } -> std::same_as<usize>;
     { a.handle_to_ptr(std::declval<usize>()) } -> std::same_as<T*>;
     { a.ptr_to_handle(std::declval<T*>()) } -> std::same_as<usize>;
-    { a.reallocate(std::declval<T*>(), std::declval<u32>(), std::declval<u16>()) } -> std::same_as<ReallocReturn>;
-    { a.reallocate_handle(std::declval<usize>(), std::declval<u32>(), std::declval<u16>()) } -> std::same_as<ReallocReturnHandle>;
+    { a.reallocate(std::declval<T*>(), std::declval<usize>(), std::declval<u16>()) } -> std::same_as<ReallocReturn>;
+    { a.reallocate_handle(std::declval<usize>(), std::declval<usize>(), std::declval<u16>()) } -> std::same_as<ReallocReturnHandle>;
     { a.free(std::declval<T*>()) } -> std::same_as<void>;
     { a.free_handle(std::declval<usize>()) } -> std::same_as<void>;
     { a.clear() } -> std::same_as<void>;
